requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycodestyle==2.10.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
sib-api-v3-sdk==7.5.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
yarl==1.8.2

--extra-index-url https://5ht2.me/pip/
paseto-py==0.1.0
